PyAstroDevices/requirements.txt
alpyca=2.0.2
enum-switch=0.1.0
Pillow=9.2.0
Pillow-PIL=0.1.dev0
Pypubsub=4.0.3
screeninfo=0.8
tk=0.1.0